Logueo de usuarios funcionando
Se eliminan los DAO y el userBean para dejar solo la estructura Facade.
