+++ TiendaJEE/web/header.xhtml
@@ -43,7 +43,7 @@ and open the template in the editor.
           </ul>
         </li>
         <h:form>
-        <div class="center-align"><h:commandButton action="#{userBean.logout}" value="Cerrar sesión" class="waves-effect waves-light btn signup red"></h:commandButton></div>
+            <div class="center-align"><h:commandButton action="#{userController.logout()}" value="Cerrar sesión" class="waves-effect waves-light btn signup red"></h:commandButton></div>
         </h:form>
     </ul>
 </ui:composition>
