+++ TiendaJEE/web/signup.xhtml
@@ -29,22 +29,22 @@
                     <h:form>
                       <div class="row">
                         <div class="input-field col s6">
-                          <h:inputText id="name" value="#{userBean.name}" class="validate"></h:inputText>
+                            <h:inputText id="name" value="#{userController.selected.name}" class="validate"></h:inputText>
                           <label for="first_name">Nombres y apellidos</label>
                         </div>
                         <div class="input-field col s6">
-                          <h:inputText id="username" value="#{userBean.username}" class="validate"></h:inputText>
+                            <h:inputText id="username" value="#{userController.selected.username}" class="validate"></h:inputText>
                           <label for="last_name">Usuario</label>
                         </div>
                       </div>
                       <div class="row">
                         <div class="input-field col s12">
-                          <h:inputSecret id="password" value="#{userBean.password}" class="validate"></h:inputSecret>
+                            <h:inputSecret id="password" value="#{userController.selected.password}" class="validate"></h:inputSecret>
                           <label for="password">Contraseña</label>
                         </div>
                       </div>
                       <div class="row">
-                        <h:commandButton action="#{userBean.signup}" value="Crear cuenta" class="waves-effect waves-light btn-large signup"></h:commandButton>
+                          <h:commandButton action="#{userController.createPublic()}" value="Crear cuenta" class="waves-effect waves-light btn-large signup"></h:commandButton>
                       </div>
                     </h:form>
                 </div>
